requirements.txt
tinkerforge
discord.py
vizdoom
pillow
numpy
opencv-python
